+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Changed
 
+- Added checks to ensure storage account `Kind` supports file and blobs before
+  attempting to fetch file and blob resources.
 - Improved error messaging when `Directory.Read.All` permission is not present
   and `config.ingestActiveDirectory=true`.
 - Handle errors thrown from `getBlobServiceProperties()` instead of allowing